cni: Convert existing flannel chaining to new chaining API
